isolatye the show/hiding method
